# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kubernetes REST API",
    "version": "1.0.0",
    "description": "A simplified REST API for managing Kubernetes resources using client-go. Supports pods, deployments, services, namespaces, configmaps, persistent volume claims, nodes, and cluster info."
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
  "paths": {
...
    "/pods": {
      "get": {
        "summary": "List pods",
        "description": "Returns a summary list of all pods. Optionally filter by namespace.",
        "operationId": "listPods",
        "tags": ["Pods"],
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Filter by namespace. Omit to list across all namespaces.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pod summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PodSummary" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/pods/{namespace}/{name}": {
      "get": {
        "summary": "Inspect a pod",
        "description": "Returns full Kubernetes Pod object.",
        "operationId": "getPod",
        "tags": ["Pods"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "200": {
            "description": "Pod object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PodSummary" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a pod",
        "description": "Deletes a pod. If managed by a controller the pod will be recreated. Pass ?force=true to set gracePeriodSeconds=0.",
        "operationId": "deletePod",
        "tags": ["Pods"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" },
          {
            "name": "force",
            "in": "query",
            "description": "Set to 'true' or '1' to delete with grace period 0.",
            "schema": { "type": "string", "enum": ["true", "1"] }
          }
        ],
        "responses": {
          "204": { "description": "Pod deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/pods/{namespace}/{name}/logs": {
      "get": {
        "summary": "Get pod logs",
        "description": "Streams the log output of a pod container as plain text.",
        "operationId": "getPodLogs",
        "tags": ["Pods"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" },
          {
            "name": "container",
            "in": "query",
            "description": "Container name (required for multi-container pods).",
            "schema": { "type": "string" }
          },
          {
            "name": "tail",
            "in": "query",
            "description": "Number of lines to return from the end of the log (default 100).",
            "schema": { "type": "integer", "default": 100 }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Only return logs after this RFC3339 timestamp.",
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Log output",
            "content": {
              "text/plain": {
                "schema": { "type": "string" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/pods/{namespace}/{name}/restart": {
      "post": {
        "summary": "Restart a pod",
        "description": "Deletes the pod so its controller (e.g. Deployment) recreates it.",
        "operationId": "restartPod",
        "tags": ["Pods"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "200": {
            "description": "Pod deleted, will be recreated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/pvcs": {
      "get": {
        "summary": "List PersistentVolumeClaims",
        "operationId": "listPVCs",
        "tags": ["PersistentVolumeClaims"],
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of PVC summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PVCSummary" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create a PersistentVolumeClaim",
        "operationId": "createPVC",
        "tags": ["PersistentVolumeClaims"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePVCRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PVC created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PVCSummary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/pvcs/{namespace}/{name}": {
      "get": {
        "summary": "Inspect a PersistentVolumeClaim",
        "operationId": "getPVC",
        "tags": ["PersistentVolumeClaims"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "200": {
            "description": "PVC object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PVCSummary" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a PersistentVolumeClaim",
        "operationId": "deletePVC",
        "tags": ["PersistentVolumeClaims"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "204": { "description": "PVC deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "namespace": {
        "name": "namespace",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Kubernetes namespace"
      },
      "name": {
        "name": "name",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Resource name"
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or missing required fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalError": {
        "description": "Kubernetes API error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "pod not found" }
        }
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "rolling restart initiated" }
        }
      },
...
      "ContainerStatus": {
        "type": "object",
        "properties": {
          "name":     { "type": "string" },
          "image":    { "type": "string" },
          "ready":    { "type": "boolean" },
          "restarts": { "type": "integer" },
          "state":    { "type": "string", "example": "running" }
        }
      },
      "PodSummary": {
        "type": "object",
        "properties": {
          "name":       { "type": "string" },
          "namespace":  { "type": "string" },
          "phase":      { "type": "string", "example": "Running" },
          "node_name":  { "type": "string" },
          "pod_ip":     { "type": "string" },
          "start_time": { "type": "string", "format": "date-time" },
          "labels":     { "type": "object", "additionalProperties": { "type": "string" } },
          "containers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ContainerStatus" }
          }
        }
      },
...
      "PVCSummary": {
        "type": "object",
        "properties": {
          "name":              { "type": "string" },
          "namespace":         { "type": "string" },
          "status":            { "type": "string", "example": "Bound" },
          "storage_class":     { "type": "string" },
          "access_modes":      { "type": "array", "items": { "type": "string" } },
          "capacity":          { "type": "string", "example": "10Gi" },
          "requested_storage": { "type": "string", "example": "10Gi" },
          "volume_name":       { "type": "string" },
          "labels":            { "type": "object", "additionalProperties": { "type": "string" } },
          "created_at":        { "type": "string", "format": "date-time" }
        }
      },
      "CreatePVCRequest": {
        "type": "object",
        "required": ["name", "storage", "access_modes"],
        "properties": {
          "name":          { "type": "string" },
          "namespace":     { "type": "string", "default": "default" },
          "storage_class": { "type": "string" },
          "access_modes":  {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ReadWriteOnce", "ReadOnlyMany", "ReadWriteMany", "ReadWriteOncePod"]
            },
            "example": ["ReadWriteOnce"]
          },
          "storage": { "type": "string", "example": "10Gi" },
          "labels":  { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
...
    }
  },
...
}